requirements.txt
fastapi
uvicorn
pymongo
pydantic
pydantic-settings
python-dotenv
google-generativeai